Move serde implementation to new module and remove unit tests
