+++ Cargo.toml
@@ -14,7 +14,7 @@ version = "1.0.0"
 [target.'cfg(all(target_family = "wasm", target_os = "unknown"))'.dependencies]
 js-sys = "0.3.20"
 wasm-bindgen = { version = "0.2.70", default-features = false }
-serde = { version = "1.0", optional = true }
+serde = { version = "1", optional = true }
 
 [dev-dependencies]
 static_assertions = "1"
